Update VGRC-VGOODRCF4.config
For the old design hardware, there are some issue, so we stopped and not sell to the market, only some test samples flashed with that version  by ourself. We will use this new updated version only. For this new version, we adjusted the I/O ports and default configs, such like LED light.etc.